requirements.txt
Django==1.8.2
Fabric==1.10.2
Markdown==2.6.2
MySQL-python==1.2.5
django-allauth==0.23.0
django-cors-headers==1.1.0
django-extensions==1.5.5
django-filter==0.10.0
django-rest-auth==0.5.0
djangorestframework==3.1.3
ecdsa==0.13
gnureadline==6.3.3
ipython==3.2.0
oauthlib==1.0.3
paramiko==1.15.2
pycrypto==2.6.1
python-openid==2.2.5
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
wsgiref==0.1.2
